Revert "Refactor/fetch error db"
