+++ backend/handlers/handleError.js
@@ -1,9 +1,8 @@
-export default function handleError(err, req, res, next) {
+export default function handleError(err, req, res) {
   console.error(err.stack);
   if (err.status) {
     res.status(err.status).json({ error: err.message });
   } else {
     res.status(500).json({ error: "Internal Server Error" });
   }
-  next();
 }
