+++ backend/server.js
@@ -30,7 +30,8 @@ const initDatabase = async () => {
         image VARCHAR(255),
         video_id VARCHAR(255),
         lesson_start_time INTEGER,
-        lesson_end_time INTEGER,        
+        lesson_end_time INTEGER,
+        audio_file TEXT,
         created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
         updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
       );
@@ -47,8 +48,7 @@ const initDatabase = async () => {
         assigned_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
         completed_at TIMESTAMP,
         updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        UNIQUE(student_id, lesson_id),
-        audio_file TEXT
+        UNIQUE(student_id, lesson_id)
       );
     `);
