+++ frontend/package.json
@@ -16,7 +16,6 @@
     "next": "15.3.4",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
-    "react-error-boundary": "^6.0.0",
     "react-h5-audio-player": "^3.10.0",
     "react-icons": "^5.5.0",
     "react-player": "^3.3.1",
