--- frontend/src/app/components/StudentLessons.module.css
@@ -1,18 +1,0 @@
-.cards-container {
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-  width: 100%;
-  min-height: max-content;
-}
-
-.skeleton-card {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  height: 100%;
-  background-color: #f0f0f0;
-  border-radius: 8px;
-  animation: pulse 1.5s infinite;
-}
