--- frontend/src/app/components/Students.module.css
@@ -1,56 +1,0 @@
-.studentsGrid {
-  display: flex;
-  flex-direction: column;
-  gap: 15px;
-}
-
-.studentCard {
-  background: white;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-  padding: 0.5rem;
-  text-decoration: none;
-  color: black;
-  box-shadow: 0 4px 20px rgba(0, 0, 0, 0.05);
-  width: 100%;
-  height: 100px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.studentCard h3 {
-  margin: 0;
-  font-size: x-large;
-  font-weight: 600;
-  color: black;
-}
-
-.button {
-  margin: 0;
-  color: white;
-  font-size: 1rem;
-  font-weight: bold;
-  background-color: #2563eb;
-  border-radius: 4px;
-  text-align: center;
-  padding: 8px 12px;
-  width: 100px;
-}
-
-@media (max-width: 768px) {
-  .studentCard {
-    padding: 0.5rem;
-    height: 70px;
-  }
-
-  .studentCard h3 {
-    font-size: large;
-  }
-
-  .button {
-    font-size: 0.8rem;
-    padding: 6px 10px;
-    width: 80px;
-  }
-}
